Error checking in objectQuery, trying to load intermediate clouds in preprocess
ObjectQuery checks that both descriptors clouds have the same type, K can be
specified in the launch file.

Attempting to load the intermediate clouds in preprocess. Needs some tweaking
because the xml parser uses push_back, so if all the clouds are not present the
number of the cloud refers to the incorrect location in the vector of clouds.