Dirty hack. I feel horrible
C++ regex is hateful and I don't want to spend more time with it - tried to find a way to pass a regex to match annotation pcd files but not normal files, and couldn't. Instead, just reading in all the matching files and making sure that no normal or feature files are read by matching in the code instead.